backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.1",
    "cryptography>=46.0.4",
    "fastapi>=0.128.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
]
